Disable 64bit tests properly
