+++ toml/toml_test.go
@@ -1,7 +1,6 @@
 package toml_test
 
 import (
-	"reflect"
 	"testing"
 
 	"github.com/influxdb/influxdb/toml"
@@ -13,20 +12,6 @@ func TestSize_UnmarshalText_MB(t *testing.T) {
 	if err := s.UnmarshalText([]byte("200m")); err != nil {
 		t.Fatalf("unexpected error: %s", err)
 	} else if s != 200*(1<<20) {
-		t.Fatalf("unexpected size: %d", s)
-	}
-}
-
-// Ensure that gigabyte sizes can be parsed.
-func TestSize_UnmarshalText_GB(t *testing.T) {
-	if typ := reflect.TypeOf(0); typ.Size() != 8 {
-		t.Skip("large gigabyte parsing on 64-bit arch only")
-	}
-
-	var s toml.Size
-	if err := s.UnmarshalText([]byte("10g")); err != nil {
-		t.Fatalf("unexpected error: %s", err)
-	} else if s != 10*(1<<30) {
 		t.Fatalf("unexpected size: %d", s)
 	}
 }
